Fix console error "may not load or link"
Fix console error saying the Netflix website "may not load or link to
moz-extension:...". The Netflix website was still using its own aes.js
and sha.js instead of the copies included in the addon.

Add aes.js and sha.js to "web_accessible_resources" so the Netflix
website will use the aes.js and sha.js included with the addon, and
avoid the error.